fix(Grid) Remove divider from docs
