+++ docs/app/Examples/collections/Grid/ResponsiveVariations/GridReversedComputerExample.js
@@ -1,0 +1,23 @@
+import React from 'react'
+import { Grid } from 'stardust'
+
+const { Column, Row } = Grid
+
+const GridReversedComputerExample = () => (
+    <Grid reversed='computer' columns='equal'>
+      <Row>
+        <Column>Computer A Fourth</Column>
+        <Column>Computer A Third</Column>
+        <Column>Computer A Second</Column>
+        <Column>Computer A First</Column>
+      </Row>
+      <Row>
+        <Column>Computer B Fourth</Column>
+        <Column>Computer B Third</Column>
+        <Column>Computer B Second</Column>
+        <Column>Computer B First</Column>
+      </Row>
+    </Grid>
+)
+
+export default GridReversedComputerExample
