+++ docs/app/Examples/collections/Grid/ResponsiveVariations/GridReversedComputerVerticallyExample.js
@@ -1,0 +1,23 @@
+import React from 'react'
+import { Grid } from 'stardust'
+
+const { Column, Row } = Grid
+
+const GridReversedMobileVerticallyExamples = () => (
+  <Grid reversed='computer vertically'>
+    <Row>
+      <Column>Computer Row 4</Column>
+    </Row>
+    <Row>
+      <Column>Computer Row 3</Column>
+    </Row>
+    <Row>
+      <Column>Computer Row 2</Column>
+    </Row>
+    <Row>
+      <Column>Computer Row 1</Column>
+    </Row>
+  </Grid>
+)
+
+export default GridReversedMobileVerticallyExamples
