+++ docs/app/Examples/collections/Grid/Types/GridExample.js
@@ -2,7 +2,7 @@ import _ from 'lodash'
 import React from 'react'
 import { Image, Grid } from 'stardust'
 
-const columns = _.range(0, 16).map((i) => (
+const columns = _.times(16, (i) => (
   <Grid.Column key={i}>
     <Image src='http://semantic-ui.com/images/wireframe/media-paragraph.png' />
   </Grid.Column>
