+++ docs/app/Examples/collections/Grid/Variations/GridPaddedHorizontallyExample.js
@@ -1,0 +1,18 @@
+import React from 'react'
+import { Image, Grid } from 'stardust'
+
+const { Column } = Grid
+const image = <Image src='http://semantic-ui.com/images/wireframe/paragraph.png' />
+
+const GridPaddedHorizontallyExample = () => (
+  <div>
+    <p>The following grid has horizontal gutters</p>
+
+    <Grid columns={2} padded='horizontally'>
+      <Column>{image}</Column>
+      <Column>{image}</Column>
+    </Grid>
+  </div>
+)
+
+export default GridPaddedHorizontallyExample
