+++ docs/app/Examples/collections/Grid/Variations/GridVerticalAlignmentGridExample.js
@@ -1,0 +1,21 @@
+import React from 'react'
+import { Grid, Image } from 'stardust'
+
+const { Column, Row } = Grid
+const image = <Image src='http://semantic-ui.com/images/wireframe/image.png' />
+
+const GridVerticalAlignmentGridExample = () => (
+    <Grid verticalAlign='middle' columns={4} centered>
+      <Row>
+        <Column>{image}</Column>
+        <Column>
+          {image}
+          <br />
+          {image}
+        </Column>
+        <Column>{image}</Column>
+      </Row>
+    </Grid>
+)
+
+export default GridVerticalAlignmentGridExample
